Rewrote deleteValueByNode, added deleteChild function, added test

+++ Dictionary/Dictionary/dictionary.c
@@ -3,9 +3,11 @@
 #include <string.h>
 #include <stdbool.h>
 
+#define MAX_LENGTH 100
+
 typedef struct {
 	int key;
-	char* value;
+	char* value[MAX_LENGTH];
 	struct Node* leftChild;
 	struct Node* rightChild;
 } Node;
@@ -18,7 +20,7 @@ Dictionary* createDictionary() {
 	return calloc(1, sizeof(Dictionary));
 }
 
-Node* leftMostChildByNode(Node* node, int key) {
+Node* leftMostChildByNode(Node* node) {
 	while (node->leftChild != NULL) {
 		node = node->leftChild;
 	}
@@ -26,42 +28,103 @@ Node* leftMostChildByNode(Node* node, int key) {
 	return node;
 }
 
-Node* leftMostChild(int key, Dictionary* dictionary) {
-	return leftMostChildByNode(dictionary->root, key);
+Node* leftMostChild(Dictionary* dictionary) {
+	return leftMostChildByNode(dictionary->root);
 }
 
-Node* deleteValueByNode(Node* node, int key) {
-	if (node == NULL) {
-		return node;
+void deleteChild(Node* parent, bool isLeft) {
+	Node* oldChild = NULL;
+	if (isLeft)
+		oldChild = parent->leftChild;
+	else
+		oldChild = parent->rightChild;
+
+	if (oldChild->leftChild == NULL && oldChild->rightChild == NULL) {
+		if (isLeft)
+			parent->leftChild = NULL;
+		else
+			parent->rightChild = NULL;
+		free(oldChild);
+		return;
 	}
 
-	if (key < node->key) {
-		deleteValueByNode(node->leftChild, key);
-	}
-	else if (key > node->key) {
-		deleteValueByNode(node->rightChild, key);
-	}
-	else if (node->leftChild != NULL && node->rightChild != NULL) {
-		Node* rightChild = node->rightChild;
-		Node* newNode = leftMostChildByNode(node, rightChild->key);
-		node->key = newNode->key;
-		node->rightChild = deleteValueByNode(node->rightChild, node->key);
-	}
-	else if (node->leftChild != NULL) {
-		node = node->leftChild;
-	}
-	else if (node->rightChild != NULL) {
-		node = node->rightChild;
+	if (oldChild->rightChild != NULL) {
+		Node* newChild = oldChild->rightChild;
+		if (newChild->leftChild == NULL) {
+			oldChild->key = newChild->key;
+			strcpy(oldChild->value, newChild->value);
+			oldChild->rightChild = newChild->rightChild;
+		}
+		else {
+			Node* parentOfNewChild = newChild;
+
+			do {
+				parentOfNewChild = newChild;
+				newChild = newChild->leftChild;
+			} while (newChild->leftChild != NULL);
+			parentOfNewChild->leftChild = newChild->rightChild;
+
+			oldChild->key = newChild->key;
+			strcpy(oldChild->value, newChild->value);
+		}
+
+		free(newChild);
 	}
 	else {
-		node = NULL;
+		Node* newChild = oldChild->leftChild;
+		oldChild->key = newChild->key;
+		strcpy(oldChild->value, newChild->value);
+		oldChild->rightChild = newChild->rightChild;
+		oldChild->leftChild = newChild->leftChild;
+		free(newChild);
+	}
+}
+
+int* deleteValueByNode(Node* node, int key) {
+	if (node->leftChild == NULL && node->rightChild == NULL)
+		return NULL;
+	if (node->key > key) {
+		if (node->leftChild == NULL)
+			return NULL;
+
+		Node* leftChild = node->leftChild;
+		if (leftChild->key == key) {
+			deleteChild(node, true);
+			return 0;
+		}
+
+		return deleteValueByNode(leftChild, key);
 	}
 
-	return node;
+	if (node->key < key) {
+		if (node->rightChild == NULL)
+			return NULL;
+
+		Node* rightChild = node->rightChild;
+		if (rightChild->key == key) {
+			deleteChild(node, false);
+			return 0;
+		}
+
+		return deleteValueByNode(rightChild, key);
+	}
 }
 
 void deleteValue(Dictionary* dictionary, int key) {
-	deleteValueByNode(dictionary->root, key);
+	Node* root = dictionary->root;
+	if (root->key == key) {
+		Node * fakeParent = calloc(1, sizeof(Node));
+		if (fakeParent == NULL) {
+			return -1;
+		}
+
+		fakeParent->leftChild = root;
+		deleteChild(fakeParent, true);
+		dictionary->root = fakeParent->leftChild;
+		free(fakeParent);
+	}
+	else
+		deleteValueByNode(dictionary->root, key);
 }
 
 void deleteDictionary(Dictionary* dictionary) {
@@ -85,8 +148,6 @@ int addNodeByNode(int key, char* value, Node* node) {
 
 		newNode->key = key;
 		strcpy(newNode->value, value);
-		newNode->leftChild = NULL;
-		newNode->rightChild = NULL;
 		node->leftChild = newNode;
 		return 0;
 	} else if (node->rightChild == NULL && (node->key < key)) {
@@ -97,16 +158,16 @@ int addNodeByNode(int key, char* value, Node* node) {
 
 		newNode->key = key;
 		strcpy(newNode->value, value);
-		newNode->leftChild = NULL;
-		newNode->rightChild = NULL;
 		node->rightChild = newNode;
 		return 0;
 	}
-	else if (node->leftChild = !NULL && key < node->key) {
+	else if (node->leftChild != NULL && key < node->key) {
 		addNodeByNode(key, value, node->leftChild);
+		return 0;
 	}
 	else if (node->rightChild != NULL && key > node->key) {
 		addNodeByNode(key, value, node->rightChild);
+		return 0;
 	}
 
 	strcpy(node->value, value);
@@ -114,6 +175,18 @@ int addNodeByNode(int key, char* value, Node* node) {
 }
 
 int addNode(Dictionary* dictionary, int key, char* value) {
+	if (dictionary->root == NULL) {
+		dictionary->root = calloc(1, sizeof(Node));
+		if (dictionary->root == NULL) {
+			return -1;
+		}
+	}
+	if (dictionary->root->key == NULL) {
+		dictionary->root->key = key;
+		strcpy(dictionary->root->value, value);
+		return 0;
+	}
+
 	int answer = addNodeByNode(key, value, dictionary->root);
 	if (answer < 0) {
 		deleteDictionary(dictionary);
@@ -124,15 +197,17 @@ int addNode(Dictionary* dictionary, int key, char* value) {
 }
 
 char* takeValueByNode(Node* node, int key) {
+	if (node == NULL)
+		return NULL;
+
 	if (node->key == key) {
 		return node->value;
 	}
 
-	char* answer = "";
 	if (key < node->key && node->leftChild != NULL) {
-		answer = takeValueByNode(node->leftChild, key);
+		return takeValueByNode(node->leftChild, key);
 	} else if (key > node->key && node->rightChild != NULL) {
-		answer = takeValueByNode(node->rightChild, key);
+		return takeValueByNode(node->rightChild, key);
 	}
 
 	return NULL;
@@ -142,7 +217,7 @@ char* takeValue(Dictionary* dictionary, int key) {
 	return takeValueByNode(dictionary->root, key);
 }
 
-bool isKeyInDictionaryByNode(Dictionary* dictionary, int key) {
+bool isKeyInDictionary(Dictionary* dictionary, int key) {
 	if (takeValueByNode(dictionary->root, key) != NULL) {
 		return true;
 	}
@@ -167,4 +242,27 @@ Node* parentByNode(Node* currentNode, int key) {
 
 Node* parent(Dictionary* dictionary, int key) {
 	return parentByNode(dictionary->root, key);
+}
+
+void printByNode(Node* node) {
+	if (node == NULL) {
+		return;
+	}
+
+	printf("%d - %s\n", node->key, node->value);
+	if (node->leftChild == NULL && node->rightChild == NULL) {
+		return;
+	}
+
+	if (node->leftChild != NULL) {
+		printByNode(node->leftChild);
+	}
+
+	if (node->rightChild != NULL) {
+		printByNode(node->rightChild);
+	}
+}
+
+void print(Dictionary* dictionary) {
+	printByNode(dictionary->root);
 }
